fix 32bit app key issue
git-svn-id: https://plugins.svn.wordpress.org/launchkey/trunk@749370 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ launchkey.php
@@ -38,7 +38,7 @@ class LaunchKey {
     public function check_option($input) {
         if(isset($input['app_key'])) {
             if(is_numeric($input['app_key']) || $input['app_key'] === '') {
-                $app_key = (int) trim($input['app_key']);
+                $app_key = trim($input['app_key']);
                 if(get_option('launchkey_app_key') === FALSE) {
                     add_option('launchkey_app_key', $app_key);
                 } else {
